Moved Orbits to its own class and simplified merge
I'm not yet sure how merging orbits of automorphisms should
be handled. Right now, two linked BitSets are just merged but
possible further automorphisms are not found. This leads to more
searches in the search tree but also to faster orbit merging. If
we consider that typical molecules have only small automorphism
groups, this might be the more efficient approach.